+++ src/Delivery/Asset.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @copyright 2015 Contentful GmbH
+ * @copyright 2015-2017 Contentful GmbH
  * @license   MIT
  */
 
@@ -9,17 +9,17 @@ namespace Contentful\Delivery;
 class Asset extends LocalizedResource implements \JsonSerializable
 {
     /**
-     * @var object
+     * @var array
      */
     private $title;
 
     /**
-     * @var object
+     * @var array
      */
     private $description;
 
     /**
-     * @var object
+     * @var array
      */
     private $file;
 
@@ -31,9 +31,9 @@ class Asset extends LocalizedResource implements \JsonSerializable
     /**
      * Asset constructor.
      *
-     * @param object           $title
-     * @param object           $description
-     * @param object           $file
+     * @param array            $title
+     * @param array            $description
+     * @param array            $file
      * @param SystemProperties $sys
      */
     public function __construct($title, $description, $file, SystemProperties $sys)
@@ -51,19 +51,30 @@ class Asset extends LocalizedResource implements \JsonSerializable
      *
      * @param  Locale|string|null $locale
      *
-     * @return string
+     * @return string|null
+     *
+     * @throws \InvalidArgumentException When $locale is not one of the locales supported by the space.
      */
     public function getTitle($locale = null)
     {
         $localeCode = $this->getLocaleFromInput($locale);
 
-        return $this->title->$localeCode;
+        // This checks happens after the call to getLocaleFromInput to make sure the Exception for invalid locales is still thrown.
+        if ($this->title === null) {
+            return null;
+        }
+
+        $localeCode = $this->loopThroughFallbackChain($this->title, $localeCode, $this->getSpace());
+
+        return $localeCode === null ? null : $this->title[$localeCode];
     }
 
     /**
      * @param  Locale|string|null $locale
      *
-     * @return string
+     * @return string|null
+     *
+     * @throws \InvalidArgumentException When $locale is not one of the locales supported by the space.
      */
     public function getDescription($locale = null)
     {
@@ -74,19 +85,23 @@ class Asset extends LocalizedResource implements \JsonSerializable
             return null;
         }
 
-        return $this->description->$localeCode;
+        $localeCode = $this->loopThroughFallbackChain($this->description, $localeCode, $this->getSpace());
+
+        return $localeCode === null ? null : $this->description[$localeCode];
     }
 
     /**
      * @param  Locale|string|null $locale
      *
      * @return File
+     *
+     * @throws \InvalidArgumentException When $locale is not one of the locales supported by the space.
      */
     public function getFile($locale = null)
     {
         $localeCode = $this->getLocaleFromInput($locale);
 
-        return $this->file->$localeCode;
+        return $this->file[$localeCode];
     }
 
     /**
@@ -160,16 +175,32 @@ class Asset extends LocalizedResource implements \JsonSerializable
      */
     public function jsonSerialize()
     {
+        $entryLocale = $this->sys->getLocale();
+
         $obj = (object) [
-            'fields' => (object) [
-                'title' => $this->title,
-                'file' => $this->file
-            ],
+            'fields' => (object) [],
             'sys' => $this->sys
         ];
+        if ($entryLocale) {
+            $obj->fields->file = $this->file[$entryLocale];
+        } else {
+            $obj->fields->file = $this->file;
+        }
+
+        if ($this->title !== null) {
+            if ($entryLocale) {
+                $obj->fields->title = $this->title[$entryLocale];
+            } else {
+                $obj->fields->title = $this->title;
+            }
+        }
 
         if ($this->description !== null) {
-            $obj->fields->description = $this->description;
+            if ($entryLocale) {
+                $obj->fields->description = $this->description[$entryLocale];
+            } else {
+                $obj->fields->description = $this->description;
+            }
         }
 
         return $obj;