+++ tests/E2E/AssetBasicTest.php
@@ -1,12 +1,13 @@
 <?php
 /**
- * @copyright 2015-2016 Contentful GmbH
+ * @copyright 2015-2017 Contentful GmbH
  * @license   MIT
  */
 
 namespace Contentful\Tests\E2E;
 
 use Contentful\Delivery\Client;
+use Contentful\Delivery\Query;
 use Contentful\ResourceArray;
 use Contentful\Delivery\Asset;
 
@@ -23,9 +24,21 @@ class AssetBasicTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @vcr e2e_asset_get_all.json
+     * @vcr e2e_asset_get_all_locale_all.json
      */
     public function testGetAll()
+    {
+        $query = (new Query())
+            ->setLocale('*');
+        $assets = $this->client->getAssets($query);
+
+        $this->assertInstanceOf(ResourceArray::class, $assets);
+    }
+
+    /**
+     * @vcr e2e_asset_get_all_locale_default.json
+     */
+    public function testGetAllSingleLocale()
     {
         $assets = $this->client->getAssets();
 
@@ -33,9 +46,20 @@ class AssetBasicTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @vcr e2e_asset_get_one.json
+     * @vcr e2e_asset_get_one_locale_all.json
      */
     public function testGetOne()
+    {
+        $asset = $this->client->getAsset('nyancat', '*');
+
+        $this->assertInstanceOf(Asset::class, $asset);
+        $this->assertEquals('nyancat', $asset->getId());
+    }
+
+    /**
+     * @vcr e2e_asset_get_one_locale_default.json
+     */
+    public function testGetOneSingleLocale()
     {
         $asset = $this->client->getAsset('nyancat');