+++ tests/E2E/GzipEncodingTest.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * @copyright 2016 Contentful GmbH
+ * @copyright 2016-2017 Contentful GmbH
  * @license   MIT
  */
 
@@ -8,6 +8,7 @@ namespace Contentful\Tests\E2E;
 
 use Contentful\Delivery\Client;
 use Contentful\Log\ArrayLogger;
+use Contentful\Delivery\Query;
 
 class GzipEncodingTest extends \PHPUnit_Framework_TestCase
 {
@@ -19,7 +20,9 @@ class GzipEncodingTest extends \PHPUnit_Framework_TestCase
         $logger = new ArrayLogger;
         $client = new Client('b4c0n73n7fu1', 'cfexampleapi', false, $logger);
 
-        $client->getEntries();
+        $query = (new Query())
+            ->setLocale('*');
+        $client->getEntries($query);
 
         $logEntry = $logger->getLogs()[0];